+++ OrderManagement.Domain/Entities/SubElement.cs
@@ -6,9 +6,9 @@
 
         public string Type { get; set; }
 
-        public float Width { get; set; }
+        public int Width { get; set; }
 
-        public float Height { get; set; }
+        public int Height { get; set; }
 
         public int WindowId { get; set; }
